+++ src/entities/AOEProjectile.java
@@ -3,7 +3,7 @@ package entities;
 import attacks.Attack;
 
 public class AOEProjectile extends Projectile{
-	public AOEProjectile(int x, int y, int dirNum, int momentum, Player attackUser, Attack a, Player chainedFrom){
+	public AOEProjectile(int x, int y, double dirNum, int momentum, Player attackUser, Attack a, Player chainedFrom){
 		super(x, y, dirNum, momentum, attackUser, a);
 		avoid(chainedFrom);
 		attackUser.getTeam().registerProjectile(this);
