+++ src/entities/Projectile.java
@@ -14,7 +14,7 @@ public class Projectile extends Entity{
 	private boolean shouldTerminate;
 	private boolean terminated;
 	
-	public Projectile(int x, int y, int dirNum, int momentum, Player attackUser, Attack a){
+	public Projectile(int x, int y, double dirNum, int momentum, Player attackUser, Attack a){
 		super(x, y, dirNum, momentum);
 		setMoving(true);
 		distanceTraveled = 0;
